Added Reservation entry to the NavigationComponent on the frontend. (#17)
* Added Reservation entry to NavigationComponent on the frontend

* Added Reservation entry to the NavigationComponent

* Removed Top-bar and Equipment components in frontend

For this specific PR, only the Reservation entry was added to the NavigationComponent on the frontend so the users can click on the feature in the sidebar. The `Reservation` entry is only available once the user is logged in using the Authentication. If the user is not logged in, only `Sign in` is displayed on the Home page.